pyproject.toml
[project]
name = "scrapefruit"
version = "0.1.0"
description = "Desktop scraping platform with GUI"
authors = [{ name = "Joe Amditis", email = "jamditis@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Desktop GUI
    "pywebview>=5.0",

    # Web framework
    "flask>=3.0",
    "flask-socketio>=5.3",
    "flask-cors>=4.0",
    "python-socketio>=5.10",
    "eventlet>=0.35",

    # Scraping
    "playwright>=1.40",
    "playwright-stealth>=1.0.6",
    "requests>=2.31",
    "trafilatura>=1.6",
    "lxml>=5.0",
    "cssselect>=1.2",

    # Database
    "sqlalchemy>=2.0",

    # Export
    "gspread>=6.0",
    "google-auth>=2.0",

    # Utils
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scrapefruit"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]
